fixed libft

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: hroussea <hroussea@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/23 14:40:04 by hroussea          #+#    #+#             */
-/*   Updated: 2020/11/26 15:30:33 by hroussea         ###   ########lyon.fr   */
+/*   Updated: 2020/12/03 16:53:20 by hroussea         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(char const *str)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (*str++)
